Updating kmeans, ssca2, and vacation to be build-able using the C++ STM API. This primarily involves extending lib/tm.h with a new set of macro defs for the CXXTM build, and pushing the [[transaction_callable]] property through the functions that are used transactionally (by adding TM_CALLABLE).
A second, more substantial change is the use of Intel's __transaction [[waiver]] to call uninstrumented comparator code, as is done in the existing STAMP design. We would like people to be able to experiment with the cost of not using waiver, so we allow users to configure it with cmake.