partial checkin: removed TxThread parameter to STM library functions, favoring instead the use of TLS.  This has a pretty big performance penalty, but appears to be how GCC does things, so if we're going with GCC, this makes sense.
This change entailed trivializing almost every metadata constructor, and then adding 'init()' methods to do the actual initialization; this is necessary if the descriptor is to be reached directly via TLS, rather than via a pointer.

Note, too, that the repeated use of 'Self' within the code does not appear to be an issue: GCC optimizes it out.  Still, it might not be a bad idea to clean all uses of Self to ensure proper caching.  I started this for NOrec, but not for anything else.