Removed a small difference between the benchmark API and STAMP API, so that we can include the benchmark API into the STAMP tm.h file.
Updated the build rules for STAMP so that we can build it using g++ and RSTM,
without the need to use TM_ARG parameters.  Only Genome and Vacation are
ported so far; the rest should be easy.

Killed off the stamp-0.9.10 folder, since 0.9.10_patched is working nicely so
far.

Updated genome and vacation so that they do not assume that lib is on the
include path.