Semi-adaptive framework implemented.  We can't dynamically change algorithms, but we can pick an algorithm at runtime.
The main value of this implementation is in showing just how much adaptivity
is going to cost.  We can compare to the single library builds and see just
how painful adaptivity is going to be.  It's bad.

There are a few temp changes in here, such as making some symbols in the
OrecLazy, OrecEager, and NOrec generic namespaces weak.  The main addition is
that now there is an adaptm target that has two-level indirection to reach
instrumentation: the API still exposes tm_XYZ functions, but they call across
a function pointer to the real code that needs to run.

Naturally CGL is penalized the worst.  But everything suffers greatly.