Introduced a generic instrumentation layer that can be used to customize read (and soon write) instrumentation for various configurations. It is implemented as clean-but-also-messy template voodoo. The voodoo is subject to change.
The upshot is that each TM that wants to use the voodoo implements intrinsic
alg_tm_read/write_aligned_word functions that encode the specific behavior they
want. Then, they instantiate the read<>/write<> template with the stack-filter
and read-after-write policies that they want, along with a parameter that
forces all access to be interpreted as aligned if desired, and the compiler
builds the appropriate instrumentation, using the intrinsic, aligned word
instrumentation.

As an example, the library API for NOrec instantiates read as 'read<T=void*,
StackFilter=NoFilter, RAW=WordlogRAW, ForceAligned=true>'. I verified that,
while the alt_tm_read asm is slightly different than when we use the previous
version, the execution time---and in particular the # branches and # branch
misses for the Tree benchmark---remains about the same.

Using this generic infrastructure, the standard ITM read barriers can be
instantiated as 'read<T=ITM_TYPE, StackFilter=FullFilter, RAW=BytelogRAW,
ForceAligned=false>'.

Also note that the 'inst-alignment.hpp' functionality automatically generates
aligned instrumentation when the architecture only supports aligned accesses
for the type, T.

While more complicated than necessary for simply implementing the ITM ABI, this
complexity allows us to selectively enable/disable features and thus measure
the cost of supporting things like byte or unaligned accesses, etc.

The 'byte-logging.hpp' header implements some macros that I hope to eliminate
in the future.

The next steps are 1) to port the rest of the implementations to use the
'inst.hpp' infrastructure, 2) to implement the write<> template, and 3) to
(hopefully) push the read-instrumentation readlogging code into the read<>
template rather than requiring it to be part of alg_tm_read_aligned_word.