Using a revised version of a generic write barrier for all of the complex TMs. The write barrier is different than the read barrier because of the danger of in-place writes for subword or unaligned accesses, where it isn't safe to just go and perform a full-word in-place write.
The current solution is to switch to using a version of the algorithm-specific
write barrier that is templated to perform a full-word write when we're not
byte-logging (or for the library API), and to perform a masked write we are.

Overall, the generic instrumentation code has turned out to be more complex
than I had hoped.