Implemented inline ASM version of transaction begin.  The new begin currently only works for the gcc_linux_ia32_opt_adapt configuration.  It *is* correct wrt adaptivity: we use a CAS to guarantee that we mark ourselves transactional before reading the function pointer.  However, there are two issues.  First, we could probably be faster with an xchg, especially now that we are doing custom ASM.  Second, the implementation does not currently update nontransactional work correctly.  I haven't decided how to handle it yet: the tick() call in stm::begin causes noticable latency, and 'nontx time' has never been a particularly useful metric for adaptivity.  Perhaps with our new algorithms, nontx time can be removed entirely?
Note: this checkin also adds a gcc_linux_amd64_opt_adapt build target, which
currently fails due to bugs in the ASM.  Consider the new .mk file a starting
point for adding amd64 inline ASM, rather than an indication that such
support is complete.

Also, note that unfortunately, I failed to move the tmbegin function pointer
out of the TxThread object in an earlier checkin, and thus this checkin has
an unnecessarily large footprint as I had to move the pointer out in order to
get reasonably straightforward name mangling for the function pointer in the
custom ASM.

Finally, while this point is documented in the code, note that the new ASM
uses parameterless functions.  There is a chance that we would save a bit by
using a parameter to pass the TxThread object, rather than accessing it via
TLS in every function.