Intermediate checkin of an alternative interface to instrumentation.
Note: this code compiles and works, but it is fickle.  You must be sure to
set STM_CONFIG accordingly, even though a single algorithm is hard-coded.

The idea behind this change is that now we can either build with per-thread
function pointers, or with global, statically linked functions.  Clearly in
the latter case, there is no adaptivity: we don't have ro and rw versions of
functions, and we even shut off the ability to adapt between algorithms.

Longer term, we'll want to expand this, so that we can have function pointers
that aren't thread-local (but of course which sacrifice ro/rw adaptivity) as
well as switch-based adaptivity without function pointers.  Note that these
features are not advocated as release-worthy, but instead are intended to
help analyze overheads, since that is a focus in the wenjia branch right now.

The code is *very* ugly right now.  We will certainly need to clean it up a
bit on account of this change.