warning: partial checkin
This checkin moves the registration functions for algorithms into macros, and gives us a mechanism for auto-building versions of algorithms that *don't* have fine-grained adaptivity, and thus do *not* require a TLS hit to access function pointers.  However, it does not currently build for the coarse adaptivity mechanism... it will soon, but there are some small bugs in terms of how the mode of the transaction is encapsulated.

Note that unless you turn on the nofga option, everything will work.